+++ source/engine/cyanomega/musepack/src/replaygain/gain_analysis.c
@@ -264,7 +264,7 @@ gain_analyze_samples ( const Float_t* left_samples, const Float_t* right_samples
         return GAIN_ANALYSIS_OK;
 
     cursamplepos = 0;
-    batchsamples = num_samples;
+    batchsamples = (long)num_samples;
 
     switch ( num_channels) {
     case  1: right_samples = left_samples;
